refactors topic menu with form components
